render a beam of light above the laser block
apparently, to render a few translucent colored faces that do NOT have untinted clouds showing through them (e.g. clouds are the same color as always, without being tinted as if they're seen through, well, a colored surface) we need quite a lot of code.
maybe it would be doable without a custom render layer and a shader program for it, but that'd require the mod to supply unused fields like texture coordinates or light levels.

TODO: generate the laser beam mesh procedurally (with configurable length, maybe rotating the ends, multiple segments, etc)